Remove obsolete optimisation
This kept breaking things. It could be made to work (I got a massive
performance gain on larger datasets on a previous project). But my
current use-case is on a comparitively tiny data where it's unlikely to
make ANY difference, let alone a noticeable one. Add to that that it is
code that will only get executed occasonally, it doesn't seem worth it.