Individual events for both primary and secondary
I originally left this off because I was thinking "well, you can have a
primary by itself, but not a secondary." But a secondary is useful for
if you want to use the secondary separately from the primary. Eg I'm
currently using the secondary to private control of CTRL, ALT, and Shift
keys.