Fix button and zone matching from macros
While I was matching hard-coded strings it worked. But once I started
taking input from macros it broke. The reason was that
    if ("string" == "string")
is actually comparing two objects are the same object, not the contents
of the strings. This is fine for hard-coded strings, but it doesn't work
when the strings come from a function that pulls the string from another
string. Using .equals() like this:
    if (input.equals("string"))
resolves this. But may come at a performance penalty. Especially since
this is a regularly executed loop. Luckily with only ~30-ish loops per
second, this has not been visible in my testing.