Rewind now stablises the cursor as well
Before, it used to leave any accumulated motion, and original location
intact. This meant that after the rewind action had completed, the
cursor would continue as it was as if nothing had happend, which wasn't
really intuitive. So now the rewind resets those things so that the
cursor continues life from the rewind point.