Use newHands support to improve gesture events
If we are in a freeze where we can't make changes, then we shouldn't be
tracking changes that are probably errors. This way when we unfreeze, we
should trigger the correct changes from our previous state rather than
the error states that may have happened in between that we weren't
interested in to begin with.