Don't allow zone override during eventFreeze
zone override is used for scrolling. Events are the way where you'd
typically enter/exit the zone override for scrolling. This meant that
during the the newHand eventFreeze, you'd continue scrolluing from
moving your hand while that was almost always not what you want to do.

Instead we now don't allow the override during that freeze, and allow it
to resume when the freeze is over. The result is no accidental scrolls
when accidentally exiting and re-entering the cone of silence. And
barely noticeable mouse movement (I really, really have to try to abuse
it to make it happen.)