Fix bug where I used the incorrect event
This is a massive facepalm. It shouldn't matter now that I have fixed
event ordering, but still enter and exit events should always be
preferred over matching state events like open and closed. The reason
for this is that you have much higher certainty that the enter/exit
events will be triggered in the order that you expect, while that might
not be true for oposing states.