Change protection behaviour to match expectations
The original behavior seemed like a good idea at the time, but it wasn't
what is generally expected from a modern UI. So I've changed it to
match.

Essentially, if a state change had gone missing, it would get undone
before being redone. When in reality, we just wanted to undo it. Eg:

You click and drag. But when you let go; the change gets missed. At this
point you click and let go again. The thing you want to happen is for
the extra down to be ignored, and the up to be captured. What was
actually happening was that an extra up was being triggered before
another full down and up cycle. The result was that the selection got
lost.

Now that the previous bug is fixed, this should not be something that we
see on a regular basis (hopefully not at all), but if we do, it should
be much less annoying that it was before.