Fix bug from copy/paste error
At some point I had pasted some random text into the command that
stablises the click when clicking on something.

I haven't looked to see how long this has been broken for, but the other
error-correction/stablisation has got so good that I didn't even notice
that this wasn't working. But when I fixed it, I definitely notice that
the clicks were sharper.