Fix grab recognition
On my device, the left two fingers are often getting stuck. The second
is the one that is being used for the grab gesture. Since the other two
fingers have not been getting stuck in my testing, this might be enough
to fix it. If not, I'll look at other solutions.

I've added a configuration item for this in ultraMotion.yml so that if
this assumption breaks for someone else, they can choose a different
finger.

Note: That this is working around a bug in LeapMotion's interpretation
of what it's seeing, not a bug with handWavey.