Add autoTrim
This trims the hand roll to account for the user being inconsistent with
their roll input. A quick motion will get past the trim and the event
will be triggered. But a slow/unintentional movement will get caught by
the autoTrim.

Trim is adjusted against the center of the current segment.