Make state changes more responsive on OOB change
Normally, state changes are stabilised to reduce flickering between
states. With this change, that stablisation is removed when the
hands enter and exit the field of view. This is so that it behaves in an
expected way when the data stops, and later restarts. Most notably:

Mouse up/Key up wasn't happening.