Make auto-trim behave better with merged segments
Previously auto-trim would treat each segment in a merged block of
segments as if it was still an individual segment. This caused it to
trim in unintuitive ways, which lead to erratic slingshotting when
coming out of a merged segment.

This change works by treating the merged segments as a whole, and only
acting on the edges of the total block.